requirements.txt
docling
pydantic
sqlalchemy
faiss-cpu
openai
python-dotenv
python-telegram-bot